+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+* Fix: add support for 2nd tuner of Hauppauge WinTV-dualHD usb tuner stick (Windows).
+
 ## Version 9.1.6 (2017-08-10)
 * Fix: Various fixes and cleanup on Linux Firewire and DVB.
 * Fix: Added support for all 4 tuners on the Hauppauge WinTV-quadHD tuner in Windows.